fix(instance-create) remove magic numbers from css and use generic way to set max-height of a component
